refactor into handlers package
